requirements.txt
psutil
requests

## MQTT
paho-mqtt
certifi

## Websockets
autobahn
twisted
pyopenssl
service_identity

## ARTIK Cloud SDK
# Uncomment to use latest released version
# artikcloud
# Uncomment to use latest development version from GitHub
#-e git+https://github.com/artikcloud/artikcloud-python.git#egg=artikcloud
